requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
filelock==3.16.1
Flask==3.0.2
flatbuffers==24.3.25
fsspec==2024.12.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.1
h5py==3.12.1
huggingface-hub==0.27.0
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.2
keras==3.7.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.0
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.2
pillow==10.2.0
protobuf==5.29.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
scikit-learn==1.5.0
scipy==1.14.1
six==1.17.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.47.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.17.0
